201509202224
git-tfs-id: [https://infovision.visualstudio.com/DefaultCollection]$/RoughSets;C225

+++ RoughSets/Infovision.Datamining.Experimenter/Parms/ParameterNumericRange.cs
@@ -31,7 +31,7 @@ namespace Infovision.Datamining.Experimenter.Parms
                     || type == typeof(short)
                     || type == typeof(byte)
                     || type == typeof(long)
-                    || type == typeof(decimal)
+                    || type == typeof(double)
                     || type == typeof(float)
                     || type == typeof(uint)
                     || type == typeof(ushort)
